test: relocate claude_cli registry tests into CI-collected tests/
The 9 registry-integration tests added by this PR (TestRunAgentRegistry,
TestReapReturncode, TestTerminateAll) lived in src/worca/utils/test_claude_cli.py,
which CI never collects: testpaths=["tests"] and the CI job runs `pytest tests/`,
so that file's tests only run when invoked by explicit path. The PR's core
regression coverage for the spawn/track/reap wiring was therefore green locally
but dead in CI.

Move the three classes verbatim into tests/test_claude_cli_registry.py (a
self-contained module, mirroring the existing test_runner_lifecycle.py /
test_runner_control_polling.py split-by-concern convention) and drop the now-unused
`subprocess` / `terminate_all` imports from the src file. All 9 tests pass under
`pytest tests/` and the unit-test layer stays green.

Pre-existing TestRunAgent failures in the src file (unmocked _resolve_tool_args
reaching subprocess.run) are unchanged and out of scope — they fail identically
on the parent commit.
